+++ src/utils/ExtensionUtils.js
@@ -31,7 +31,8 @@
 define(function (require, exports, module) {
     "use strict";
     
-    var FileSystem = require("filesystem/FileSystem");
+    var FileSystem = require("filesystem/FileSystem"),
+        FileUtils  = require("file/FileUtils");
 
     /**
      * Appends a <style> tag to the document's head.
@@ -217,6 +218,31 @@ define(function (require, exports, module) {
         return result.promise();
     }
     
+    /**
+     * Loads the package.json file in the given extension folder.
+     *
+     * @param {string} folder The extension folder.
+     * @return {$.Promise} A promise object that is resolved with the parsed contents of the package.json file,
+     *     or rejected if there is no package.json or the contents are not valid JSON.
+     */
+    function loadPackageJson(folder) {
+        var file = FileSystem.getFileForPath(folder + "/package.json"),
+            result = new $.Deferred();
+        FileUtils.readAsText(file)
+            .done(function (text) {
+                try {
+                    var json = JSON.parse(text);
+                    result.resolve(json);
+                } catch (e) {
+                    result.reject();
+                }
+            })
+            .fail(function () {
+                result.reject();
+            });
+        return result.promise();
+    }
+    
     exports.addEmbeddedStyleSheet = addEmbeddedStyleSheet;
     exports.addLinkedStyleSheet   = addLinkedStyleSheet;
     exports.parseLessCode         = parseLessCode;
@@ -224,4 +250,5 @@ define(function (require, exports, module) {
     exports.getModuleUrl          = getModuleUrl;
     exports.loadFile              = loadFile;
     exports.loadStyleSheet        = loadStyleSheet;
+    exports.loadPackageJson       = loadPackageJson;
 });
